apps/api/pyproject.toml
[project]
name = "finetune-api"
version = "0.1.0"
description = "Online finetune platform API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "bcrypt>=4.0.0",
  "fastapi>=0.116.0",
  "uvicorn>=0.35.0",
  "pydantic>=2.11.0",
  "dependency-injector>=4.46.0",
  "omegaconf>=2.3.0",
  "httpx>=0.28.0",
  "prefect>=3.0.0",
  "label-studio-sdk>=1.0.0",
  "sqlalchemy>=2.0.36",
  "aiosqlite>=0.20.0",
  "asyncpg>=0.30.0",
  "psycopg2-binary>=2.9.10",
  "alembic>=1.14.1",
  "minio>=7.2.14",
  "kubernetes>=31.0.0",
  "python-multipart>=0.0.9",
  "python-jose[cryptography]>=3.3.0",
  "pgvector>=0.3.0",
  "grpcio>=1.68.0",
  "embedding-protos",
  "croniter>=2.0.0",
  "pillow>=11.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.24.0",
]

[tool.uv.sources]
embedding-protos = { workspace = true }

[tool.pytest.ini_options]
testpaths = ["tests"]
faulthandler_timeout = 60
faulthandler_exit_on_timeout = true
markers = [
    "no_auth_override: skip the global auth dependency override for this test",
]
